+++ spring-cloud-starter-tencent-metadata-transfer/src/main/java/com/tencent/cloud/metadata/core/EncodeTransferMedataRestTemplateInterceptor.java
@@ -21,6 +21,7 @@ package com.tencent.cloud.metadata.core;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
+import java.util.HashMap;
 import java.util.Map;
 
 import com.tencent.cloud.common.constant.MetadataConstant;
@@ -33,9 +34,12 @@ import org.springframework.http.HttpRequest;
 import org.springframework.http.client.ClientHttpRequestExecution;
 import org.springframework.http.client.ClientHttpRequestInterceptor;
 import org.springframework.http.client.ClientHttpResponse;
+import org.springframework.lang.NonNull;
 import org.springframework.util.CollectionUtils;
 
 import static com.tencent.cloud.common.constant.ContextConstant.UTF_8;
+import static com.tencent.cloud.common.constant.MetadataConstant.HeaderName.CUSTOM_DISPOSABLE_METADATA;
+import static com.tencent.cloud.common.constant.MetadataConstant.HeaderName.CUSTOM_METADATA;
 
 /**
  * Interceptor used for adding the metadata in http headers from context when web client
@@ -51,22 +55,44 @@ public class EncodeTransferMedataRestTemplateInterceptor implements ClientHttpRe
 	}
 
 	@Override
-	public ClientHttpResponse intercept(HttpRequest httpRequest, byte[] bytes,
-			ClientHttpRequestExecution clientHttpRequestExecution) throws IOException {
+	public ClientHttpResponse intercept(@NonNull HttpRequest httpRequest, @NonNull byte[] bytes,
+			@NonNull ClientHttpRequestExecution clientHttpRequestExecution) throws IOException {
 		// get metadata of current thread
 		MetadataContext metadataContext = MetadataContextHolder.get();
 		Map<String, String> customMetadata = metadataContext.getFragmentContext(MetadataContext.FRAGMENT_TRANSITIVE);
+		Map<String, String> disposableMetadata = metadataContext.getFragmentContext(MetadataContext.FRAGMENT_DISPOSABLE);
 
-		if (!CollectionUtils.isEmpty(customMetadata)) {
-			String encodedTransitiveMetadata = JacksonUtils.serialize2Json(customMetadata);
+		Map<String, String> newestCustomMetadata = new HashMap<>();
+		customMetadata.forEach((key, value) -> {
+			if (!disposableMetadata.containsKey(key)) {
+				newestCustomMetadata.put(key, value);
+			}
+		});
+		// build custom disposable metadata request header
+		this.buildMetadataHeader(httpRequest, disposableMetadata, CUSTOM_DISPOSABLE_METADATA);
+
+		// build custom metadata request header
+		this.buildMetadataHeader(httpRequest, newestCustomMetadata, CUSTOM_METADATA);
+
+		return clientHttpRequestExecution.execute(httpRequest, bytes);
+	}
+
+	/**
+	 * Set metadata into the request header for {@link HttpRequest} .
+	 *
+	 * @param request    instance of {@link HttpRequest}
+	 * @param metadata   metadata map .
+	 * @param headerName target metadata http header name .
+	 */
+	private void buildMetadataHeader(HttpRequest request, Map<String, String> metadata, String headerName) {
+		if (!CollectionUtils.isEmpty(metadata)) {
+			String encodedMetadata = JacksonUtils.serialize2Json(metadata);
 			try {
-				httpRequest.getHeaders().set(MetadataConstant.HeaderName.CUSTOM_METADATA,
-						URLEncoder.encode(encodedTransitiveMetadata, UTF_8));
+				request.getHeaders().set(headerName, URLEncoder.encode(encodedMetadata, UTF_8));
 			}
 			catch (UnsupportedEncodingException e) {
-				httpRequest.getHeaders().set(MetadataConstant.HeaderName.CUSTOM_METADATA, encodedTransitiveMetadata);
+				request.getHeaders().set(headerName, encodedMetadata);
 			}
 		}
-		return clientHttpRequestExecution.execute(httpRequest, bytes);
 	}
 }