+++ CHANGELOG.md
@@ -8,3 +8,8 @@
 - [feat:enhance Feign and RestTemplate and support Polaris monitor.](https://github.com/Tencent/spring-cloud-tencent/pull/447)
 - [Feature: support spring cloud config data.](https://github.com/Tencent/spring-cloud-tencent/pull/451)
 - [Optimize feign & rest-template circuit-breaker logic](https://github.com/Tencent/spring-cloud-tencent/pull/454)
+- [Feature: Add disposable metadata transfer support](https://github.com/Tencent/spring-cloud-tencent/pull/459)
+- [docs:update mvnw.](https://github.com/Tencent/spring-cloud-tencent/pull/476)
+- [docs:update configuration metadata.](https://github.com/Tencent/spring-cloud-tencent/pull/474)
+- [Feature: delete implement ServiceInstance](https://github.com/Tencent/spring-cloud-tencent/pull/483)
+- [test: add loadbalancer unit test](https://github.com/Tencent/spring-cloud-tencent/pull/485)